+++ DBTesting/DBTesting/DB/Models/UserEntity.cs
